+++ CHANGELOG.md
@@ -15,10 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The disconnected method on the `portal` directive now checks that the portal container is in the target's children.
   If not, then a warning will be printed that the container was already removed from the target.
 
-  The prevalence of this issue is unknown, but it was encountered while working with components in portals rendered using `open-wc`'s `fixture` helper.
+  The prevalence of this issue is unknown, but it was encountered while working with components
+  in portals rendered using `open-wc`'s `fixture` helper.
+- Refactor lambdas out of tests to support setting the Mocha timeout limit.
 
 ### Added
 - Tests for Lit lifecycle methods called on components that are in portals.
+- Support for asynchronous portal content, with tests and demo code.
 
 ## [0.6.2] - 2024-04-09